add NotificationSettingsRepository, NotificationTypeRepository, putApiAccountNotifications and change error code to 400 in AccountController

+++ src/main/java/ru/skillbox/socialnetwork/api/responses/NotificationBaseResponse.java
@@ -11,10 +11,10 @@ public class NotificationBaseResponse {
 
   private long id;
   @JsonProperty("type_id")
-  private int typeId;
+  private long typeId;
   @JsonProperty("sent_time")
   private long sentTime;
   @JsonProperty("entity_id")
-  private int entityId;
+  private long entityId;
   private String info;
 }
